requirements.txt
pyyaml>=6.0
requests>=2.31.0
psutil>=5.9.0
requests[socks]>=2.31.0
pysocks>=1.7.1
